<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

Currently this is very WIP - the first commit straight up deletes `StarPrerouting` so I could work incrementally while still having something build.

This PR does really a lot too much; probably I'll split off many small changes before the main PR, and it also changes more about the Sabre internals than I really _meant_ to change as part of this - I kind of mixed together making a few changes (though they are related).

From a pure-Rust interface perspective, there are several things going on here, ticks indicate whether they're done or not:

- [x] Move all descriptions of the routing target information and construction of them into Rust (the new `RoutingTarget`)
- [x] Move all construction of the Sabre interaction DAG to Rust (the new `SabreDAG`)
- [x] Move `SabreSwap` DAG rebuild to Rust (the new `RoutingResult` and `RoutingResult::rebuild`)
- [x] Rework Python and Rust entry points `sabre_routing`/`swap_map` to use the new structures and returns.
- [x] Rewrite Python `SabreSwap` to use the new Rust components.
- [ ] Move `SabreLayout` DAG rebuild to Rust (`RoutingResult::rebuild_onto` is most of the rebuild, but it's not hooked up).
- [ ] Rewrite Python-space `SabreLayout` to use the new Rust components.
- [ ] Fix `StarPrerouting`.

As of commit 815bc3d95f8c4aa768068f3530d8594ced98afbb (the original HEAD of this PR), `SabreSwap` works (I believe) completely correctly end-to-end.  `SabreLayout` is currently totally broken in Python space because I haven't finished rewriting the pass yet.  `StarPrerouting` is straight-up deleted because it reaches deep into Sabre internals, and I didn't want to have to deal with the pass just to have a working build - it'll be restored at some point.

### Details and comments

Close #12279 
Close #12280

#### New `SabreDAG`

The new `SabreDAG` is far more detached from the Python-space one, and needed (or perhaps just _wanted_) its control-flow handling rewritten so the ownership model both of `SabreDAG` and the output `RoutingResult` was far clearer when it came to control-flow blocks.  Previously, the ownership around the temporary `DAGCircuit`s of the inner blocks was quite awkward in the Rust/Python split, with the blocks being flattened in the result, but Python space holding onto the temporary `DAGCircuit`s backing the blocks, in order to do the rebuild later.  The new `SabreDAG` stores the temporary `DAGCircuit`s within itself, next to the `SabreDAG` that refers to that block.  This part of the change made the move to pure-Rust result and rebuild operations much more natural and easier.  The types of node in the graph are now called `Synchronize`, `TwoQ` and `ControlFlow`, where `Synchronize` is an operation that has no routing requirements, it only imposes data-synchronisation constraints (like a barrier, say).

An additional change to `SabreDAG` that probably _shouldn't_ be in this PR (and I might separate out later) is the coalescence of runs of operations onto the same arguments, or subsets, into the same node in the Sabre interaction graph, wherever possible.  For example, any 1q gate is compressed into its preceding 2q gate.  Similarly, any time we see a 2q gate followed by zero or more 1q gates on those two qubits, then another 2q gate on the same two qubits, the second 2q gate imposes no possible further routing constraints, so is always compressed into the prior node.  I mostly made this change as a performance optimisation for `update_route` and `populate_extended_set` (there are fewer nodes to explicitly visit), and because I wanted to avoid the `HashMap`-based `SwapMap` strategy of the Python-space result object, and instead have the swaps inline.  The cost, though, is that I didn't want the result object to have loads of empty `Vec`s in it - coalescing the nodes in the interaction graph has the knock-on effect of making it so _most_ of the individual nodes in the result will have initial swaps.

#### New `RoutingResult`

This replaces the tuple of a few `pyclass`es, in favour of single unified Rust handling.  The `RoutingResult` object stores a reference to the `SabreDAG` (and the `DAGCircuit`) it relates to, so it can rebuild itself, and so that it needn't store copies of the full routing order; just as the prior Python result object stored node indices in the Python DAG to facilitate the rebuild, this one stores node indices in the `SabreDAG`, so we get the full order of coalesced nodes without cloning it into the result object.

The new `RoutingResult` object can completely reconstruct the physical output when acting in `sabre_routing` mode, or (in theory - untested as yet) `sabre_layout` mode when there is no disjoint handling required.  `RoutingResult::rebuild` is effectively a full implementation of circuit embedding, layout application and routing rebuild.  Going further, `RoutingResult::rebuild_onto` _should_ (again, untested) work to manage rebuilding onto a larger `DAGCircuit` after routing on only a subset of the `Target` (such as happens in disjoint-target handling).

#### New `RoutingTarget`

The previous `(NeighborTable, Graph<(), ()>, ArrayView2<f64>)` form of the target specification was strongly influenced by what we could create and pass around from Python, but also had a fair amount of duplication in it.  Previously we had the problem that the Rustworkx Python-exposed `CouplingMap` couldn't be consumed by our Rust passes, so we were creating a `NeighborTable` from a list of edges, _rebuilding_ a `Graph<(), ()>` from that, and calculating a distance matrix entirely separately.

Now we're Rust-native, we only want to work with a `Target` as input.

In the new form: the roles fulfilled by `NeighborTable` and `Graph<(), ()>` are combined into a single new `Neighbors` structure, which then implements all the `petgraph::visit` traits using `PhysicalQubit` as the index type, so can be passed natively to `dijkstra` (for the release valve) and `token_swapper` (for control-flow layout reversion), and in turn, the result swap paths come out already in terms of `PhysicalQubit`.

Internally, `Neighbors` is a flat list `[n0_0, n0_1, n1_0, n1_1, n1_2, n2_0, ...]`, where `nX_Y` means the `Y`th neighbor of qubit `X`, and a separate list of partition points into that flat list.  Fundamentally, it is the same as the list-of-lists approach of the old `NeighborTable`, just flattened so we don't even need `SmallVec` and its "did this spill to the heap?" runtime check on each access for data locality; the data is simply always localised because it is flat and contiguous.  It is similar to the CSR-like structure used by `SparseObservable`, for example.  `Neighbors` then implements all the suitable petgraph visitor traits as if it were an undirected graph.  Actually, it's more like a guaranteed-symmetric directed graph (each edge is duplicated) for performance reasons in swap-candidate enumeration with how Sabre works, but that largely doesn't matter.

The distance matrix is now calculated directly from `Neighbors`.  The relevant method isn't in `rustworkx_core` yet (see Qiskit/rustworkx_core#1439), so this patch contains a (re-optimised) port of that method.