@jakelishman 2025-05-13
As of (commit hash will change in future rebases)
```
3de0913cc8 Ensure ancillas are allocated in sorted order
```
the tests (other than the disabled disjoint ones) should all pass.  Following commits affect the randomisation, and will be split off into separate PRs.