@jakelishman 2025-05-20
As noted in the now-updated PR comment: there's still some minor work to fix up the last couple of `todo!` stubs, though I think that mostly they're still `todo!` because I wasn't immediately certain how they should be handled, and we don't have test-suite coverage of them.

I also haven't run benchmarks on this any time recently.  From a now-old run, I know that one of the ASV utility-scale benchmarks shows a regression, but I tracked it down to a miniscule deviation in (I think) floating-point stuff 10k+ steps into the Sabre routing algorithm.  Functionally, the only reason it appears at all is because #14244 is not fixed by this PR, and the dodgy extended-set tracking causes the heuristic to be _completely_ borked for huge circuits.

I have a couple of PRs to follow this one that break RNG compatibility.  They build on stuff that might seem a bit superfluous in the current PR - the reason is because I wrote way too much at once, and then modified the patch series back into an RNG-compatible patch followed by improvements, so some of the extra structures appear to get added before they're properly taken advantage of, but it was too hard to write another completely new implementation to get rid of them.