@jakelishman 2025-05-22
Local benchmarking suggests that while this has expected performance improvements in `sabre_routing`, there's some sort of regression in `sabre_layout` that needs more investigation.