@jakelishman 2025-05-23
No longer "on hold" for performance: the trouble was that this commit forcibly disabled double parallelisation in the routing component of a single layout trial, and on machines where the single layer of parallelism didn't saturate the thread count, that was a regression.  The intent had been to reduce parallelism synchronisation overhead, but in practice, each routing run takes long enough, and Rayon is efficient enough under these circumstances, that it's irrelevant.