@jakelishman 2025-05-23
Changed benchmarks, running on the 192-thread benchmarking server.

```text
| Change   | Before [22367fca]    | After [a102ff63]    |   Ratio | Benchmark (Parameter)                                                                                   |
|----------|----------------------|---------------------|---------|---------------------------------------------------------------------------------------------------------|
| -        | 2.78±0.04ms          | 2.49±0.06ms         |    0.9  | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 131072)                             |
| -        | 2.80±0.03ms          | 2.52±0.01ms         |    0.9  | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 131072)                             |
| -        | 2.80±0.03ms          | 2.49±0.06ms         |    0.89 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 131072)                              |
| -        | 2.80±0.02ms          | 2.38±0.05ms         |    0.85 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 131072)                              |
| -        | 76.1±1ms             | 29.4±0.9ms          |    0.39 | mapping_passes.PassBenchmarks.time_sabre_swap(14, 1024)                                                 |
| -        | 123±3ms              | 47.8±1ms            |    0.39 | mapping_passes.PassBenchmarks.time_sabre_swap(20, 1024)                                                 |
| -        | 37.1±9ms             | 8.33±1ms            |    0.22 | mapping_passes.PassBenchmarks.time_sabre_swap(5, 1024)                                                  |
| -        | 37.8±0.5μs           | 34.3±0.08μs         |    0.91 | passes.PassBenchmarks.time_remove_diagonal_gates_before_measurement(20, 1024)                           |
| -        | 17.4±0.09μs          | 15.8±0.1μs          |    0.91 | passes.PassBenchmarks.time_remove_diagonal_gates_before_measurement(5, 1024)                            |
| -        | 10.1±0.02s           | 5.61±0.01s          |    0.55 | qft.LargeQFTMappingTimeBench.time_sabre_swap(1081, 'decay')                                             |
| -        | 9.76±0.01s           | 5.22±0s             |    0.53 | qft.LargeQFTMappingTimeBench.time_sabre_swap(1081, 'lookahead')                                         |
| -        | 81.7±10ms            | 33.4±0.1ms          |    0.41 | qft.LargeQFTMappingTimeBench.time_sabre_swap(115, 'decay')                                              |
| -        | 102±0.9ms            | 32.0±0.06ms         |    0.31 | qft.LargeQFTMappingTimeBench.time_sabre_swap(115, 'lookahead')                                          |
| -        | 1.23±0.02s           | 570±3ms             |    0.46 | qft.LargeQFTMappingTimeBench.time_sabre_swap(409, 'decay')                                              |
| -        | 1.23±0.01s           | 548±7ms             |    0.44 | qft.LargeQFTMappingTimeBench.time_sabre_swap(409, 'lookahead')                                          |
| -        | 5.85±0.06ms          | 5.22±0.02ms         |    0.89 | qft.QftTranspileBench.time_ibmq_backend_transpile(1)                                                    |
| -        | 6.85±0.2ms           | 5.99±0.02ms         |    0.88 | qft.QftTranspileBench.time_ibmq_backend_transpile(2)                                                    |
| -        | 2.88±0.01s           | 2.56±0.01s          |    0.89 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(1081, 10, 'lookahead')                |
| -        | 20.7±8ms             | 9.54±0.08ms         |    0.46 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(115, 10, 'decay')                     |
| -        | 20.3±0.2ms           | 9.30±0.06ms         |    0.46 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(115, 10, 'lookahead')                 |
| -        | 182±10ms             | 78.8±1ms            |    0.43 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(115, 100, 'decay')                    |
| -        | 178±8ms              | 74.7±0.7ms          |    0.42 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(115, 100, 'lookahead')                |
| -        | 276±10ms             | 192±0.2ms           |    0.69 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(409, 10, 'decay')                     |
| -        | 404±10ms             | 303±0.6ms           |    0.75 | quantum_volume.LargeQuantumVolumeMappingTimeBench.time_sabre_swap(409, 10, 'lookahead')                 |
| -        | 14.7±4ms             | 6.70±0.02ms         |    0.46 | quantum_volume.QuantumVolumeBenchmark.time_ibmq_backend_transpile(1, 'synthesis')                       |
| -        | 16.3±2ms             | 5.66±0.02ms         |    0.35 | quantum_volume.QuantumVolumeBenchmark.time_ibmq_backend_transpile(1, 'translator')                      |
| -        | 16.3±8ms             | 7.45±0.03ms         |    0.46 | quantum_volume.QuantumVolumeBenchmark.time_ibmq_backend_transpile(2, 'synthesis')                       |
| -        | 8.21±1ms             | 6.44±0.02ms         |    0.78 | quantum_volume.QuantumVolumeBenchmark.time_ibmq_backend_transpile(2, 'translator')                      |
| -        | 149±5ms              | 124±2ms             |    0.83 | quantum_volume.QuantumVolumeBenchmark.time_ibmq_backend_transpile(20, 'synthesis')                      |
| -        | 7.42±0.06ms          | 6.52±0.02ms         |    0.88 | queko.QUEKOTranspilerBench.time_transpile_bigd(0, None)                                                 |
| -        | 6.38±3ms             | 5.12±0.01ms         |    0.8  | queko.QUEKOTranspilerBench.time_transpile_bigd(1, None)                                                 |
| -        | 8.25±0.2ms           | 7.11±0.06ms         |    0.86 | queko.QUEKOTranspilerBench.time_transpile_bigd(2, None)                                                 |
| -        | 22.3±8ms             | 10.8±0.04ms         |    0.48 | queko.QUEKOTranspilerBench.time_transpile_bntf(1, None)                                                 |
| -        | 84.8±5ms             | 70.9±0.7ms          |    0.84 | queko.QUEKOTranspilerBench.time_transpile_bntf(2, 'sabre')                                              |
| -        | 19.8±0.5ms           | 17.0±0.05ms         |    0.86 | queko.QUEKOTranspilerBench.time_transpile_bntf(2, None)                                                 |
| -        | 32.7±8ms             | 29.6±0.08ms         |    0.91 | queko.QUEKOTranspilerBench.time_transpile_bntf(3, None)                                                 |
| -        | 109±10ms             | 75.6±8ms            |    0.69 | queko.QUEKOTranspilerBench.time_transpile_bss(0, 'sabre')                                               |
| -        | 74.6±10ms            | 44.9±0.5ms          |    0.6  | queko.QUEKOTranspilerBench.time_transpile_bss(0, None)                                                  |
| -        | 20.2±10ms            | 15.9±0.04ms         |    0.79 | queko.QUEKOTranspilerBench.time_transpile_bss(1, None)                                                  |
| -        | 110±4ms              | 80.3±5ms            |    0.73 | queko.QUEKOTranspilerBench.time_transpile_bss(2, 'sabre')                                               |
| -        | 55.2±1ms             | 28.4±0.09ms         |    0.52 | queko.QUEKOTranspilerBench.time_transpile_bss(2, None)                                                  |
| -        | 211±10ms             | 130±9ms             |    0.62 | queko.QUEKOTranspilerBench.time_transpile_bss(3, 'sabre')                                               |
| -        | 413                  | 323                 |    0.78 | queko.QUEKOTranspilerBench.track_depth_bss_optimal_depth_100(1, 'sabre')                                |
| -        | 293                  | 239                 |    0.82 | queko.QUEKOTranspilerBench.track_depth_bss_optimal_depth_100(3, 'sabre')                                |
| -        | 7.87±0.05ms          | 7.14±0.02ms         |    0.91 | random_circuit_hex.BenchRandomCircuitHex.time_ibmq_backend_transpile(4)                                 |
| -        | 9.49±0.07ms          | 6.66±0.05ms         |    0.7  | ripple_adder.RippleAdderTranspile.time_transpile_square_grid_ripple_adder(10, 0)                        |
| -        | 17.0±0.2ms           | 10.9±0.05ms         |    0.64 | ripple_adder.RippleAdderTranspile.time_transpile_square_grid_ripple_adder(20, 0)                        |
| -        | 105±10ms             | 44.4±5ms            |    0.42 | ripple_adder.RippleAdderTranspile.time_transpile_square_grid_ripple_adder(20, 1)                        |
| -        | 96.4±2ms             | 60.8±5ms            |    0.63 | ripple_adder.RippleAdderTranspile.time_transpile_square_grid_ripple_adder(20, 2)                        |
| -        | 86                   | 78                  |    0.91 | statepreparation.StatePreparationTranspileBench.track_cnot_counts_after_mapping_to_ibmq_16_melbourne(6) |
| -        | 187                  | 161                 |    0.86 | statepreparation.StatePreparationTranspileBench.track_cnot_counts_after_mapping_to_ibmq_16_melbourne(7) |
| -        | 394                  | 342                 |    0.87 | statepreparation.StatePreparationTranspileBench.track_cnot_counts_after_mapping_to_ibmq_16_melbourne(8) |
| -        | 19.7±9ms             | 10.00±0.1ms         |    0.51 | transpiler_benchmarks.TranspilerBenchSuite.time_compile_from_large_qasm                                 |
| -        | 7.19±0.1ms           | 5.73±0.02ms         |    0.8  | transpiler_benchmarks.TranspilerBenchSuite.time_cx_compile                                              |
| -        | 14.4±4ms             | 6.02±0.02ms         |    0.42 | transpiler_benchmarks.TranspilerBenchSuite.time_single_gate_compile                                     |
| -        | 62.9±0.2ms           | 54.2±0.2ms          |    0.86 | transpiler_levels.TranspilerLevelBenchmarks.time_quantum_volume_transpile_50_x_20(0)                    |
| -        | 159±2ms              | 137±2ms             |    0.86 | transpiler_levels.TranspilerLevelBenchmarks.time_quantum_volume_transpile_50_x_20(1)                    |
| -        | 209±4ms              | 175±2ms             |    0.84 | transpiler_levels.TranspilerLevelBenchmarks.time_quantum_volume_transpile_50_x_20(2)                    |
| -        | 42.0±8ms             | 18.3±0.07ms         |    0.44 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm(1)                           |
| -        | 30.1±7ms             | 10.9±0.04ms         |    0.36 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm(2)                           |
| -        | 24.8±10ms            | 13.3±0.02ms         |    0.54 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm(3)                           |
| -        | 160±4ms              | 119±2ms             |    0.75 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_qv_14_x_14(3)                                |
| -        | 728                  | 598                 |    0.82 | transpiler_levels.TranspilerLevelBenchmarks.track_depth_transpile_qv_14_x_14(3)                         |
| -        | 5.41±0.04ms          | 3.84±0.04ms         |    0.71 | transpiler_qualitative.TranspilerQualitativeBench.time_transpile_time_cnt3_5_179(0)                     |
| -        | 8.26±0.05ms          | 5.53±0.05ms         |    0.67 | transpiler_qualitative.TranspilerQualitativeBench.time_transpile_time_cnt3_5_180(0)                     |
| -        | 71.0±5ms             | 54.0±0.9ms          |    0.76 | transpiler_qualitative.TranspilerQualitativeBench.time_transpile_time_cnt3_5_180(2)                     |
| -        | 8.58±0.09ms          | 5.66±0.03ms         |    0.66 | transpiler_qualitative.TranspilerQualitativeBench.time_transpile_time_qft_16(0)                         |
| -        | 25.3±0.7ms           | 7.73±0.03ms         |    0.31 | transpiler_qualitative.TranspilerQualitativeBench.time_transpile_time_qft_16(2)                         |
| -        | 9.94±0.05ms          | 8.58±0.02ms         |    0.86 | transpiler_qualitative.TranspilerQualitativeBench.time_transpile_time_qft_16(3)                         |
| -        | 423±5ms              | 371±3ms             |    0.88 | utility_scale.UtilityScaleBenchmarks.time_qft('cx')                                                     |
| -        | 533±10ms             | 481±1ms             |    0.9  | utility_scale.UtilityScaleBenchmarks.time_qft('cz')                                                     |
| -        | 536±10ms             | 484±2ms             |    0.9  | utility_scale.UtilityScaleBenchmarks.time_qft('ecr')                                                    |
| Change   |   Before [22367fca]  |   After [a102ff63]  |   Ratio | Benchmark (Parameter)                                                                   |
|----------|----------------------|---------------------|---------|-----------------------------------------------------------------------------------------|
| +        |                   54 |                  68 |    1.26 | queko.QUEKOTranspilerBench.track_depth_bigd_optimal_depth_45(1, 'sabre')                |
| +        |                 1209 |                1625 |    1.34 | queko.QUEKOTranspilerBench.track_depth_bntf_optimal_depth_25(0, None)                   |
| +        |                  270 |                 352 |    1.3  | queko.QUEKOTranspilerBench.track_depth_bntf_optimal_depth_25(1, 'sabre')                |
| +        |                  682 |                 760 |    1.11 | queko.QUEKOTranspilerBench.track_depth_bss_optimal_depth_100(0, 'sabre')                |
| +        |                  250 |                 359 |    1.44 | queko.QUEKOTranspilerBench.track_depth_bss_optimal_depth_100(2, 'sabre')                |
| +        |                  660 |                 997 |    1.51 | ripple_adder.RippleAdderTranspile.track_depth_transpile_square_grid_ripple_adder(20, 3) |
| +        |                  819 |                 925 |    1.13 | transpiler_levels.TranspilerLevelBenchmarks.track_depth_transpile_qv_14_x_14(1)         |
| +        |                 2649 |                2982 |    1.13 | utility_scale.UtilityScaleBenchmarks.track_qv_depth('cx')                               |
| +        |                 2649 |                2982 |    1.13 | utility_scale.UtilityScaleBenchmarks.track_qv_depth('cz')                               |
| +        |                 2649 |                2982 |    1.13 | utility_scale.UtilityScaleBenchmarks.track_qv_depth('ecr')                              |
```